requirements.txt
config==0.5.1
fastapi==0.115.8
gunicorn==23.0.0
openai==1.64.0
pipreqs==0.4.13
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0

